moved image cropper code into its own file
